Refactor TabDetails
- Pull out functions into their own files
- Pull tests out into their own files
- Group all logic into one folder
